Replace timeout function with external dependency
Motivated by the combination of wanting to test
`tryExecuteSirenAction` and realising that we would just be testing
`timeoutPromise`. We have already decided to replace exactly the same
function in
`comit-rs` (https://github.com/comit-network/comit-rs/pull/2330#discussion_r399873316),
so replacing it here has the nice consequence of not needing to write
a new test, since it's not our own code.

Side-effects:

- Move `TryParams` to `swap.ts` since it's directly related and only
  used for executing a swap.
- Rename `util/timeout_promise.ts` to `util/sleep.ts`, since it now
  only holds a `sleep` function.